requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
google-generativeai==0.4.0
pyttsx3==2.90
edge-tts==6.1.10
elevenlabs==0.2.27
pydantic==2.5.3
